refactor into DecoratingBinder
